# packages/api-contract/openapi.yaml
openapi: 3.1.0

info:
  title: MapIt API
  version: 0.1.0
  description: |
    API de MapIt — motor de gestión espacial multi-tenant.

    **Este archivo es la fuente de verdad del contrato.** Se edita ANTES de escribir
    código, y de él se generan:
      - el cliente TypeScript que consume Angular (`libs/api-client`)
      - las interfaces Java que los `@RestController` deben implementar

    Consecuencia: si alguien cambia el contrato sin actualizar este archivo,
    el build de Java falla porque el controlador ya no implementa la interfaz.
    El contrato lo hace cumplir el compilador, no la convención.

    Regenerar con `pnpm api:gen`; verificar que no hay drift con `pnpm api:check`.

  contact:
    name: Equipo MapIt
  license:
    name: MIT
    identifier: MIT

servers:
  - url: http://localhost:8080/api/v1
    description: Desarrollo local
  - url: /api/v1
    description: Relativo (contenedor)

tags:
  - name: health
    description: Estado del servicio
  - name: feature-flags
    description: Feature toggles evaluados para el contexto actual
  - name: demo-items
    description: CRUD de demostración para validar la integración Angular + Spring Boot

paths:
  /health:
    get:
      tags: [health]
      summary: Estado del servicio
      description: Endpoint de humo. Sirve para verificar que el contrato de punta a punta funciona.
      operationId: getHealth
      security: []
      responses:
        "200":
          description: El servicio está operativo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /feature-flags:
    get:
      tags: [feature-flags]
      summary: Flags evaluadas para el usuario y tenant actuales
      description: |
        Devuelve el estado de las feature flags ya evaluadas en el contexto del
        llamante (tenant, usuario, rol, país). El frontend no evalúa reglas:
        solo consume booleanos.

        Nota: para el navegador, la vía preferente es el **unleash-proxy**
        (`UNLEASH_PROXY_URL`). Este endpoint existe para diagnóstico y para
        clientes que ya están autenticados contra el backend.
      operationId: getFeatureFlags
      responses:
        "200":
          description: Mapa de flag → habilitada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureFlags"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /demo-items:
    get:
      tags: [demo-items]
      summary: Listar elementos de demostración
      description: Endpoint temporal para probar el flujo CRUD completo sin autenticación JWT.
      operationId: listDemoItems
      security: []
      responses:
        "200":
          description: Elementos del tenant de demostración
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DemoItem"
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      tags: [demo-items]
      summary: Crear un elemento de demostración
      description: Endpoint temporal para probar el flujo CRUD completo sin autenticación JWT.
      operationId: createDemoItem
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoItemRequest"
      responses:
        "201":
          description: Elemento creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  /demo-items/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Identificador del elemento
        schema:
          type: string
          format: uuid

    get:
      tags: [demo-items]
      summary: Obtener un elemento de demostración
      description: Endpoint temporal para probar el flujo CRUD completo sin autenticación JWT.
      operationId: getDemoItem
      security: []
      responses:
        "200":
          description: Elemento solicitado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoItem"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

    put:
      tags: [demo-items]
      summary: Actualizar un elemento de demostración
      description: Endpoint temporal para probar el flujo CRUD completo sin autenticación JWT.
      operationId: updateDemoItem
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoItemRequest"
      responses:
        "200":
          description: Elemento actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

    delete:
      tags: [demo-items]
      summary: Eliminar un elemento de demostración
      description: Endpoint temporal para probar el flujo CRUD completo sin autenticación JWT.
      operationId: deleteDemoItem
      security: []
      responses:
        "204":
          description: Elemento eliminado
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT emitido por `/auth/login`. Incluye el claim `tenant`, que es de
        donde el backend resuelve el tenant (NO del header `X-Tenant-Id`, que el
        cliente podría falsificar).

  schemas:
    HealthStatus:
      type: object
      required: [status, version]
      properties:
        status:
          type: string
          enum: [UP, DOWN]
          examples: [UP]
        version:
          type: string
          examples: ["0.1.0"]
        timestamp:
          type: string
          format: date-time

    FeatureFlags:
      type: object
      description: Clave de flag (`<dominio>.<feature>`) → si está habilitada.
      additionalProperties:
        type: boolean
      examples:
        - demo.hello: true
          payments.qr: false
          realtime.websocket: true

    DemoItem:
      type: object
      required: [id, name, description, active, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true

    DemoItemRequest:
      type: object
      required: [name, description, active]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
        active:
          type: boolean

    # RFC 9457 — Problem Details. Spring 6+ lo soporta de forma nativa,
    # así que todos los errores de la API usan este formato.
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
          examples: ["Unauthorized"]
        status:
          type: integer
          examples: [401]
        detail:
          type: string
        instance:
          type: string
          format: uri

  responses:
    Unauthorized:
      description: Falta el token o no es válido
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    BadRequest:
      description: La solicitud no cumple las validaciones
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ServerError:
      description: Error inesperado del servidor
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Autenticado, pero sin permiso para este recurso o tenant
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: El recurso no existe (o pertenece a otro tenant)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

security:
  - bearerAuth: []